fix: fix lab2 divide and round function

+++ lab2/DebugExercise/DebugExercise1.java
@@ -6,7 +6,7 @@ package DebugExercise;
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+        float quotient = (float) top / bottom;
         int result = Math.round(quotient);
         return result;
     }
@@ -21,7 +21,6 @@ public class DebugExercise1 {
         int b2 = 4;
         int result2 = divideThenRound(t2, b2);
         System.out.println("round(" + t2 + "/" + b2 + ")=" + result2);
-
         int t3 = 3;
         int b3 = 4;
         int result3 = divideThenRound(t3, b3);
